feat(starkey): separate eip2645 in eth folder
