+++ src/HDWallet.cpp
@@ -11,6 +11,7 @@
 #include "Bitcoin/CashAddress.h"
 #include "Bitcoin/SegwitAddress.h"
 #include "Coin.h"
+#include "Ethereum/eip2645.h"
 #include "Mnemonic.h"
 #include "memory/memzero_wrapper.h"
 
@@ -263,6 +264,10 @@ PrivateKey HDWallet::bip32DeriveRawSeed(TWCoinType coin, const Data& seed, const
     auto data = Data(node.private_key, node.private_key + PrivateKey::_size);
     TW::memzero(&node);
     return PrivateKey(data);
+}
+
+std::string HDWallet::eip2645(const std::string& ethAddress, const std::string& layer, const std::string& application, const std::string& index) const {
+    return Ethereum::accountPathFromAddress(ethAddress, layer, application, index);
 }
 
 namespace {
