+++ src/ImmutableX/Constants.h
@@ -19,8 +19,6 @@ inline const int256_t gStarkCurveB("31415926535897932384626433832795028841971693
 inline const int256_t gStarkCurveN("3618502788666131213697322783095070105526743751716087489154079457884512865583");
 inline const int256_t gStarkCurveP("3618502788666131213697322783095070105623107215331596699973092056135872020481");
 inline const int256_t gStarkDeriveBias("112173586448650067624617006275947173271329056303198712163776463194419898833073");
-/*inline const CurvePoint gStarkCurveG{.x = int256_t("874739451078007766457464989774322083649278607533249481151382481072868806602"),
-                                     .y = int256_t("152666792071518830868575557812948353041420400780739481342941381225525861407")};*/
 } // namespace internal
 
 } // namespace TW::ImmutableX
