+++ tests/chains/ImmutableX/StarkKeyTests.cpp
@@ -12,7 +12,7 @@ namespace TW::ImmutableX::tests {
 
 TEST(ImmutableX, PathFromAddress) {
     // https://github.com/immutable/imx-core-sdk-swift/blob/main/Tests/ImmutableXCoreTests/Crypto/Stark/StarkKeyTests.swift#L30
-    auto res = accountPathFromAddress("0xa76e3eeb2f7143165618ab8feaabcd395b6fac7f");
+    auto res = Ethereum::accountPathFromAddress("0xa76e3eeb2f7143165618ab8feaabcd395b6fac7f", internal::gLayer, internal::gApplication, internal::gIndex);
     ASSERT_EQ(res, "m/2645'/579218131'/211006541'/1534045311'/1431804530'/1");
 }
